#1 UI Menu
+UI menu
+menu

+++ src/core/transaction.py
@@ -1,5 +1,5 @@
 class Transaction:
-    def __init__(self, date: str, amount: float, category: str):
+    def __init__(self, date: str, amount: int, category: str):
         self.date = date
         self.amount = amount
         self.category = category
